Refactored Employee dashboard and moved RequestStatusCard to common/
