--- src/components/Admin/RequestStatusCard/index.js
@@ -1,3 +1,0 @@
-import RequestStatusCard from './RequestStatusCard';
-
-export default RequestStatusCard;
