--- src/components/Employee/Dashboard/EmployeeDashboard.js
@@ -1,22 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  flex-direction: column;
-  background-color: #3aafa9;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 4px;
-  padding: 0 15px;
-  font-size: 15px;
-  line-height: 1;
-  font-weight: 500;
-  height: 50px;
-`;
-
-export const Title = styled.h1`
-  margin: 2px 10px;
-  font-size: 3.75rem;
-  color: #17252a;
-  font-weight: 700;
-`;
