+++ src/components/common/RequestStatusCard/index.js
@@ -1,0 +1,3 @@
+import RequestStatusCard from './RequestStatusCard.jsx';
+
+export default RequestStatusCard;
